# api-spec.yaml
openapi: 3.0.3
info:
  title: AgentHub Registry API
  description: |
    The universal package registry for AI-native agents, tools, chains, and prompts.

    ## Authentication

    This API uses GitHub OAuth 2.0 for authentication. Most endpoints require a valid JWT token.

    ## Rate Limiting

    API requests are rate limited to 1000 requests per minute per IP address.

    ## Package Types

    The registry supports the following package types:
    - `agent`: AI agents and assistants
    - `tool`: Utilities and functions for agents
    - `chain`: Workflow chains and pipelines
    - `prompt`: Prompt templates and examples
    - `dataset`: Training and reference datasets

    ## Versioning

    All packages follow semantic versioning (semver) conventions.
  version: 1.0.0
  contact:
    name: AgentHub Team
    url: https://github.com/agenthubcli/registry
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://registry.agenthubcli.com
    description: Production server
  - url: http://localhost:8000
    description: Development server
paths:
...
  /api/v1/auth/github:
    get:
      summary: Initiate GitHub OAuth
      description: Get GitHub OAuth authorization URL to start login process
      tags: [Authentication]
      parameters:
        - name: redirect_to
          in: query
          description: URL to redirect to after successful authentication
          schema:
            type: string
            format: uri
      responses:
        "200":
          description: OAuth URL generated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthUrl"
  /api/v1/auth/github/callback:
    get:
      summary: GitHub OAuth callback
      description: Handle GitHub OAuth callback and issue JWT tokens
      tags: [Authentication]
      parameters:
        - name: code
          in: query
          required: true
          description: Authorization code from GitHub
          schema:
            type: string
        - name: state
          in: query
          description: State parameter to prevent CSRF attacks
          schema:
            type: string
      responses:
        "200":
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSuccess"
        "400":
          description: Invalid authorization code or OAuth error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/auth/refresh:
    post:
      summary: Refresh access token
      description: Get a new access token using a refresh token
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          description: Invalid refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/auth/me:
    get:
      summary: Get current user
      description: Get information about the currently authenticated user
      tags: [Authentication]
      security:
        - BearerAuth: []
      responses:
        "200":
          description: User information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/auth/logout:
    post:
      summary: Logout
      description: Logout the current user (client-side token invalidation)
      tags: [Authentication]
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # Package Management Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained via GitHub OAuth
...
  schemas:
...
    Error:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
          description: Error message
        error_code:
          type: string
          description: Machine-readable error code
      example:
        detail: "Package not found"
        error_code: "PACKAGE_NOT_FOUND"
    Message:
      type: object
      required: [message]
      properties:
        message:
          type: string
          description: Success message
      example:
        message: "Operation completed successfully"

    # API Info
...
    OAuthUrl:
      type: object
      required: [oauth_url]
      properties:
        oauth_url:
          type: string
          format: uri
          description: GitHub OAuth authorization URL
          example: "https://github.com/login/oauth/authorize?client_id=..."
    AuthSuccess:
      type: object
      required: [access_token, refresh_token, token_type, user]
      properties:
        access_token:
          type: string
          description: JWT access token
        refresh_token:
          type: string
          description: JWT refresh token
        token_type:
          type: string
          example: "bearer"
        expires_in:
          type: integer
          description: Access token expiry in seconds
          example: 3600
        user:
          $ref: "#/components/schemas/UserProfile"
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
          description: JWT refresh token
    TokenResponse:
      type: object
      required: [access_token, token_type]
      properties:
        access_token:
          type: string
          description: New JWT access token
        token_type:
          type: string
          example: "bearer"
        expires_in:
          type: integer
          description: Access token expiry in seconds
          example: 3600

    # User
    UserProfile:
      type: object
      required: [id, github_username]
      properties:
        id:
          type: integer
          example: 123
        github_username:
          type: string
          example: "johndoe"
        display_name:
          type: string
          example: "John Doe"
        github_avatar_url:
          type: string
          format: uri
          example: "https://github.com/avatars/johndoe"
        bio:
          type: string
          example: "AI researcher and developer"
        website:
          type: string
          format: uri
          example: "https://johndoe.dev"
        location:
          type: string
          example: "San Francisco, CA"
        company:
          type: string
          example: "AI Company Inc."
        total_packages:
          type: integer
          example: 5
        total_downloads:
          type: integer
          example: 1234
        created_at:
          type: string
          format: date-time

    # Package Upload
...